requirements.txt
gunicorn==20.1.0
alembic==1.13.2
aniso8601==9.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.33
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.3

